Edit code snippets to replace Tailwind utility classes with CSS modules
As discussed with Johan. The React template for this guide has also been edited to remove Tailwind CSS and use CSS modules instead.

You can find this template here: https://github.com/gdangelo/cra-template-quickstart-nhost. Please make sure to move it to the Nhost Github repository and publish it to npm.