Use es2019 build target
It seems that if the target to esbuild is es2019 it transforms the
nullish coalescing (??) operator, which causes problem in React Native and
Expo. Starting from es2020 it does not transforms those.

From the documentation I thought we should use es2020, but actually we need
es2019 to make it work.

https://esbuild.docschina.org/content-types/#javascript